Changed to avoid these MS compiler messages: d:\coin\coinall-trunk\cbc\src\cbclinked.cpp(3746) : warning C4244: '=' : conversion from 'const double' to 'float', possible loss of data d:\coin\coinall-trunk\cbc\src\cbclinked.cpp(3753) : warning C4244: '=' : conversion from 'const double' to 'float', possible loss of data d:\coin\coinall-trunk\cbc\src\cbclinked.cpp(3825) : warning C4244: '=' : conversion from 'double' to 'float', possible loss of data

+++ trunk/Cbc/src/CbcLinked.hpp
@@ -333,7 +333,7 @@ private:
     */
     unsigned int type:4; // type of computation
     unsigned int affected:25; // variable or element affected
-    float multiplier; // to use in computation 
+    double multiplier; // to use in computation 
   } boundElementAction;
   
   /**@name Private member data */
